Switch from JSON to SQLite

+++ src-tauri/src/model/dto/session.rs
@@ -2,11 +2,11 @@ use serde::{Deserialize, Serialize};
 
 use super::MsqtDto;
 
-#[derive(Clone, Serialize, Deserialize, Debug)]
+#[derive(Clone, Serialize, Deserialize, Debug, sqlx::FromRow)]
 #[serde(rename_all = "camelCase")]
 pub struct Session {
-    pub(in crate::model) id: u64,
-    pub(in crate::model) fk_selected_server_id: Option<u64>,
+    pub(in crate::model) id: u32,
+    pub(in crate::model) fk_selected_server_id: Option<u32>,
     pub(in crate::model) listen_all_topics: bool,
     pub(in crate::model) connected: bool,
 }
@@ -24,14 +24,14 @@ impl Default for Session {
 }
 
 impl MsqtDto for Session {
-    fn id(&self) -> u64 {
+    fn id(&self) -> u32 {
         self.id
     }
 }
 
 #[allow(dead_code)]
 impl Session {
-    pub fn server_id(&self) -> Option<u64> {
+    pub fn server_id(&self) -> Option<u32> {
         self.fk_selected_server_id
     }
     pub fn listen_all_topics(&self) -> bool {
